put classes in separate folders
