+++ src/files.js
@@ -1,3 +1,12 @@
+/**
+ * 🔹 Logs a string to a specified log file 🔹
+ * @param {string} path Relative path to the log file
+ * @param {string} content Content that should be written to the log file
+ * @param {Object} [options] Additional options
+ * @param {boolean} [options.append_bottom=true] true to append content to the bottom of a file, false to just override the file's contents
+ * @param {boolean} [options.timestamp=false] true to add a timestamp to the logged content
+ * @since 1.5.0
+ */
 function logger(path, content, options) {
     let fs = require("fs");
     let isEmpty = require("./misc").isEmpty;
@@ -11,14 +20,15 @@ function logger(path, content, options) {
     if(!options.append_bottom) fs.writeFileSync(path, content);
     else fs.appendFileSync(path, content + "\n");
 }
-
 module.exports.logger = logger;
 
-
-
-
-
-
+/**
+ * 🔹 Reads a folder asynchronously and recursively and returns all absolute file paths (starting at the drive letter (eg. "C:/Users/...")) in the callback - Warning! Large amounts of files (like letting it run on "C:/") can freeze the process completely or exceed the maximum possible index of a JS array 🔹
+ * @param {String} folder The folder that should be recursively read
+ * @param {Function} callback The function that gets called after the folder has been read - has two arguments: error and result
+ * @async
+ * @since 1.7.0
+ */
 function readdirRecursive(folder, callback) { // modified version of https://stackoverflow.com/a/5827895/8602926
     let fs = require("fs");
     let path = require("path");
@@ -48,7 +58,11 @@ function readdirRecursive(folder, callback) { // modified version of https://sta
 }
 module.exports.readdirRecursive = readdirRecursive;
 
-
+/**
+ * 🔹 Reads a folder synchronously and recursively and returns all absolute file paths (starting at the drive letter (eg. "C:/Users/...")) in the callback - Warning! Large amounts of files (like letting it run on "C:/") can freeze the process completely or exceed the maximum possible index of a JS array 🔹
+ * @param {String} folder The folder that should be recursively read
+ * @since 1.7.0
+ */
 function readdirRecursiveSync(folder) { // from https://stackoverflow.com/a/16684530/8602926
     let fs = require("fs");
     let walk = function(dir) {
